# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HappyDrive API
  version: 1.1.0
  description: |
    日本語iOS・企業ポータル・運営Web共通契約。全書込に監査イベントを記録する。
    v1.0（docs/spec/openapi.v1.0-original.yaml）からの変更は後方互換の追加のみ。変更履歴は docs/CONTRACT_CHANGELOG.md。
    - 時刻はUTC（ISO 8601）で送受信し、表示はJST。金額は整数JPY。
    - エラーは常に Error スキーマ（message は日本語）。
    - 重要な書込は Idempotency-Key 必須。同一キー・同一内容の再送は同じ応答を返し、内容が異なれば 422 idempotency_mismatch。
servers:
  - url: https://api.happydrive.jp/v1
  - url: http://localhost:8080/v1
security:
  - bearerAuth: []
...
paths:
...
  /assignments/{assignmentId}/messages:
    get:
      tags: [Messaging]
      operationId: listMessages
      parameters: [{$ref: '#/components/parameters/AssignmentId'}, {name: after, in: query, schema: {type: string, format: date-time}}]
      responses:
        default: {$ref: '#/components/responses/Default'}
        '200': {description: 案件チャット, content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/Message'}}}}}
    post:
      tags: [Messaging]
      operationId: sendMessage
      parameters: [{$ref: '#/components/parameters/AssignmentId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [body], properties: {body: {type: string, minLength: 1, maxLength: 2000}, evidenceId: {type: string, format: uuid}}}}}}
      responses:
        default: {$ref: '#/components/responses/Default'}
        '201': {description: 送信, content: {application/json: {schema: {$ref: '#/components/schemas/Message'}}}}
        '403': {$ref: '#/components/responses/Forbidden'}
  /reports:
    post:
      tags: [Messaging]
      operationId: createReport
      description: 不適切なメッセージ・案件・利用者の通報。
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [targetType, targetId, reason], properties: {targetType: {type: string, enum: [message, job, user, organization]}, targetId: {type: string, format: uuid}, reason: {type: string, enum: [harassment, fraud, unsafe, illegal, spam, privacy, other]}, detail: {type: string, maxLength: 2000}}}}}}
      responses:
        default: {$ref: '#/components/responses/Default'}
        '201': {description: 受付, content: {application/json: {schema: {$ref: '#/components/schemas/Report'}}}}
  /blocks:
    post:
      tags: [Messaging]
      operationId: blockOrganization
      description: 発注組織をブロック。以後その組織の案件は検索・推薦に出ない。
      requestBody: {required: true, content: {application/json: {schema: {type: object, additionalProperties: false, required: [organizationId], properties: {organizationId: {type: string, format: uuid}}}}}}
      responses:
        default: {$ref: '#/components/responses/Default'}
        '204': {description: ブロック}

  # ---------------------------------------------------------------- Evidence
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
...
    AssignmentId: {name: assignmentId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
  responses:
...
    Forbidden: {description: 権限がない, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
...
    Default: {description: その他のエラー（401 認証 / 403 権限 / 404 不存在・他組織 / 422 入力 / 429 回数制限 / 5xx）。本文は常に Error, content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}}
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string, description: '機械判定用コード（例 capacity_full / not_eligible / invalid_state）'}
        message: {type: string, description: 利用者に表示できる日本語}
        requestId: {type: string}
        details: {type: object, additionalProperties: true}
...
    Message:
      type: object
      required: [id, assignmentId, senderRole, body, createdAt]
      properties: {id: {type: string, format: uuid}, assignmentId: {type: string, format: uuid}, senderRole: {type: string, enum: [worker, organization, operator, system]}, senderName: {type: string}, isMine: {type: boolean}, body: {type: string}, evidenceId: {type: string, format: uuid}, hidden: {type: boolean}, createdAt: {type: string, format: date-time}}
    Report:
      type: object
      required: [id, targetType, targetId, reason, status, createdAt]
      properties: {id: {type: string, format: uuid}, reporterId: {type: string, format: uuid}, targetType: {type: string}, targetId: {type: string, format: uuid}, reason: {type: string}, detail: {type: string}, status: {type: string, enum: [open, resolved]}, resolution: {type: string}, createdAt: {type: string, format: date-time}}
...